refactor(ui): extract message-part types to dedicated module
Slice 11b.1 step 1/14 of the message-part.tsx split mandated by
AGENTS.md (>500-line files must be split) and the slice design doc
(GPT-X-authored 14-file split plan).

Extracts the type-only declarations — `MessageProps`, `SessionAction`,
`UserActions`, `MessagePartProps`, `PartComponent`, `ToolProps`,
`ToolComponent` — into a new `message-part-types.ts` module. The
new file is a pure type module (no runtime exports, no JSX, no Solid
primitive use), so renderer/tool files can pull in the shared shapes
without dragging the heavier message-part.tsx transitive import graph.

`message-part.tsx` keeps its existing public API by re-exporting the
types verbatim, so `session-turn.tsx` / `message-part.stories.tsx` /
unit tests need no update in this step. The next slice will migrate
consumers to import from `./message-part-types` directly and retire
the re-exports.

Net: message-part.tsx loses ~35 declaration lines; types live in a
49-line file. Typecheck GREEN, 478 pass / 3 baseline fail unchanged.
